closes #41. create Vet service impl

+++ pet-clinic-data/src/main/java/com/atech/repositories/VetRepository.java
@@ -3,5 +3,5 @@ package com.atech.repositories;
 import com.atech.entity.Vet;
 import org.springframework.data.repository.CrudRepository;
 
-public interface VetRepository extends CrudRepository<Vet, Integer> {
+public interface VetRepository extends CrudRepository<Vet, Long> {
 }
